server/requirements.txt
fastapi
uvicorn
telethon
python-dotenv
opencv-python-headless
pytesseract